Projet Flask/pyproject.toml
[tool.poetry]
name = "flaskgpt"
version = "0.1.0"
description = "\"Clone\" de chatGPT avec FLASK"
authors = ["Peter"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.0"
openai = "0.27.10"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
